+++ src/logger/logger.go
@@ -7,17 +7,6 @@ import (
 	"runtime"
 )
 
-type Color string
-
-const (
-	ColorBlack  Color = "\u001b[30m"
-	ColorRed    Color = "\u001b[31m"
-	ColorGreen        = "\u001b[32m"
-	ColorYellow       = "\u001b[33m"
-	ColorBlue         = "\u001b[34m"
-	ColorReset        = "\u001b[0m"
-)
-
 func colorize(color Color, message string) {
 	fmt.Println(string(color), message, string(ColorReset))
 }
@@ -25,14 +14,9 @@ func colorize(color Color, message string) {
 func colorizeSameLine(color Color, message string) {
 	fmt.Print(string(color), message, string(ColorReset))
 }
-
-// var re = regexp.MustCompile(`^.*\.(.*)$`)
-
-// func Trim(input string) string {
-// 	res1 := strings.Trim(input, "Documents/")
-// 	return res1
-// }
-
+type Opt struct {
+	Color Color
+}
 // LogTrace logs only value with color
 func LogTrace(name string, output interface{}) {
 	// PrintUi("=", 40, false)